sistema_de_informes/services/ai-orchestrator/requirements.txt
fastapi==0.111.0
uvicorn==0.30.0
python-dotenv==1.0.1
pydantic==2.8.2
